+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Added parameter `pad` to `_.partition`
 * Added `_.sliding`
 * Added `_.sample` and `_.sampleProb`
+* Added `_.where`
 
 ##1.4.0 (07/14/14)